Overhauled Items
I don't think I'll continue working on this. It's a nice idea, but it lacks too much to be a real project. If I am going to build something like this, I shouldn't do it based on an hour and 20 minute live coding session I had in class. I should spend the time to think about what I want and actually create something meaningful and fun.

I might come back and get a minimum viable example going, but just fleshing out items shows what a monumental task it is for even the simplest of logic in something like a game.