Create Analytic_functions.sql
I added these because under the hood SQL uses hashing a lot.

[The below is not verified, but it makes sense to me. I could be WRONG]
Analytic functions are a good example, try running any of that code on PSQL and see that a new row with a sort of "Key" (maybe the better word is value) exists.

Not as well known as hash joins.